requirements.txt
discord.py
pytz
requests
gspread
oauth2client
numpy
pymongo
dnspython
googletrans
dotenv